requirements.txt
discord.py
jishaku
PyGithub